+++ .env
@@ -1,5 +1,5 @@
-SKIP_PREFLIGHT_CHECK = true;
 API_KEY: '285646332687546';
 API_SECRET: 'UAse27C9NMMXX9x1bYUvbpGxhl4';
 CLOUD_NAME: 'riyasharma';
 UPLOAD_PRESET: 'q2fmc4ar';
+SKIP_PREFLIGHT_CHECK = true;
